AdFacade created
