+++ src/main/java/ua/tc/marketplace/facade/AdFacade.java
@@ -1,0 +1,20 @@
+package ua.tc.marketplace.facade;
+
+import java.util.Map;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import ua.tc.marketplace.model.dto.ad.AdDto;
+import ua.tc.marketplace.model.dto.ad.CreateAdDto;
+import ua.tc.marketplace.model.dto.ad.UpdateAdDto;
+
+public interface AdFacade {
+  Page<AdDto> findAll(Map<String, String> filterCriteria, Pageable pageable);
+
+  AdDto findAdById(Long adId);
+
+  AdDto createNewAd(CreateAdDto dto);
+
+  AdDto updateAd(Long adId, UpdateAdDto dto);
+
+  void deleteAd(Long adId);
+}
